Code/pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.37",
]
